+++ s4e-web/src/app/views/settings/state/institution/institution.model.ts
@@ -1,6 +1,7 @@
 export interface Institution {
   id: string;
 
+  ancestorDepth?: number;
   parentName: string|null;
   parentSlug: string|null;